rename robot_trait.rs to private_robot.rs
