+++ src/robot/private_robot.rs
